Refactor investment projects collection to use form macro configs
