+++ src/apps/components/form/macros.js
@@ -1,66 +1,4 @@
-const metadata = require('../../../lib/metadata')
-const { transformObjectToOption, transformStringToOption } = require('../../transformers')
-
-const globalFields = {
-  countries: {
-    macroName: 'MultipleChoiceField',
-    name: 'country',
-    label: 'Country',
-    initialOption: '-- Select country --',
-    options () {
-      return metadata.countryOptions.map(transformObjectToOption)
-    },
-  },
-
-  sectors: {
-    macroName: 'MultipleChoiceField',
-    name: 'sector',
-    label: 'Sectors',
-    initialOption: '-- Select sector --',
-    options () {
-      return metadata.sectorOptions.map(transformObjectToOption)
-    },
-  },
-
-  strategicDrivers: {
-    macroName: 'MultipleChoiceField',
-    name: 'strategicDrivers',
-    label: 'Strategic drivers',
-    options () {
-      return metadata.strategicDriverOptions.map(transformObjectToOption)
-    },
-  },
-
-  averageSalary: {
-    macroName: 'MultipleChoiceField',
-    type: 'radio',
-    name: 'averageSalary',
-    label: 'Average salary range',
-    options () {
-      return metadata.salaryRangeOptions.map(transformObjectToOption)
-    },
-  },
-
-  foreignOtherCompany: {
-    macroName: 'MultipleChoiceField',
-    name: 'foreignOtherCompany',
-    label: 'Type of organisation',
-    initialOption: '-- Select organisation type --',
-    options () {
-      return foreignOtherCompanyOptions.map(transformStringToOption)
-    },
-  },
-}
-
-const foreignOtherCompanyOptions = [
-  'Charity',
-  'Company',
-  'Government dept',
-  'Intermediary',
-  'Limited partnership',
-  'Partnership',
-  'Sole trader',
-]
+const { globalFields } = require('../../macros')
 
 const standardFormConfig = {
   children: [
@@ -177,7 +115,6 @@ const entitySearchConfig = {
 }
 
 module.exports = {
-  globalFields,
   standardFormConfig,
   entitySearchConfig,
 }
