+++ src/apps/investment-projects/builders.js
@@ -1,8 +1,7 @@
-const { get, isString, isArray, merge } = require('lodash')
+const { get, isString, isArray } = require('lodash')
 const { transformObjectToOption } = require('../transformers')
 const metadataRepo = require('../../lib/metadata')
 const { collectionFilterLabels } = require('./labels')
-const { FILTERS_MACRO_CONFIG, SORT_OPTIONS } = require('./constants')
 
 function buildInvestmentFilters (filtersQuery = {}) {
   const formOptions = {
@@ -11,7 +10,7 @@ function buildInvestmentFilters (filtersQuery = {}) {
     sector: metadataRepo.sectorOptions.map(transformObjectToOption),
   }
 
-  const filters = Object.keys(collectionFilterLabels.edit).reduce((filtersObj, filterName) => {
+  return Object.keys(collectionFilterLabels.edit).reduce((filtersObj, filterName) => {
     const filterOptions = formOptions[filterName] || []
 
     let valueLabel = filtersQuery[filterName]
@@ -42,24 +41,8 @@ function buildInvestmentFilters (filtersQuery = {}) {
 
     return filtersObj
   }, {})
-
-  return merge(filters, FILTERS_MACRO_CONFIG)
-}
-
-function buildInvestmentSorting (filtersQuery = {}) {
-  const query = Object.assign(
-    {},
-    { sortby: SORT_OPTIONS[0].value },
-    filtersQuery
-  )
-
-  return {
-    options: SORT_OPTIONS,
-    selected: query.sortby,
-  }
 }
 
 module.exports = {
   buildInvestmentFilters,
-  buildInvestmentSorting,
 }
