--- src/apps/investment-projects/constants.js
@@ -1,37 +1,0 @@
-const FILTERS_MACRO_CONFIG = {
-  stage: {
-    macroName: 'MultipleChoiceField',
-    type: 'checkbox',
-  },
-  investment_type: {
-    macroName: 'MultipleChoiceField',
-    type: 'checkbox',
-  },
-  sector: {
-    macroName: 'MultipleChoiceField',
-    initialOption: 'All sectors',
-  },
-  estimated_land_date_before: {
-    macroName: 'TextField',
-    hint: 'YYYY-MM-DD',
-    placeholder: 'e.g. 2018-07-18',
-  },
-  estimated_land_date_after: {
-    macroName: 'TextField',
-    hint: 'YYYY-MM-DD',
-    placeholder: 'e.g. 2019-05-09',
-  },
-}
-
-const SORT_OPTIONS = [
-  { value: 'estimated_land_date:asc', label: 'Earliest land date' },
-  { value: 'estimated_land_date:desc', label: 'Latest land date' },
-  { value: 'project_code', label: 'Project code' },
-  { value: 'name:asc', label: 'Project name' },
-  { value: 'stage.name', label: 'Stage' },
-]
-
-module.exports = {
-  FILTERS_MACRO_CONFIG,
-  SORT_OPTIONS,
-}
